requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
jinja2>=3.1,<4
python-multipart>=0.0.9,<1
passlib[argon2]>=1.7,<2
argon2-cffi>=23.1,<24
itsdangerous>=2.2,<3
apscheduler>=3.10,<4
praw>=7.7,<8
python-dotenv>=1.0,<2
requests>=2.32,<3
